# Massively speed up the query returned by TI.filter_for_tis