@ashb 2020-09-25
/cc @KevinYang21 Found this from testing your huge dag -- it was called as part of `update_state`.