@ashb 2020-09-25
137 on test mysql. Will retry and see if it was just unlucky. (I tested locally to get numbers, and it runs fine.)