@potiuk 2020-09-25
I think it was unluckly. Other MySQL Core tests worked, and 137 happens occasionally